+++ lib/src/base_downloader.dart
@@ -103,9 +103,14 @@ abstract base class BaseDownloader {
         // map is <taskId, Task/progress> where progress is added to Task JSON
         final payload = progressUpdateMap[taskId];
         final task = Task.createFromJsonMap(payload);
-        final double progress = payload['progress'] ?? progressFailed;
+        final double progress = switch (payload['progress']) {
+          int a => a.toDouble(),
+          double a => a,
+          _ => progressFailed
+        };
         final int expectedFileSize = payload['expectedFileSize'] ?? -1;
-        processProgressUpdate(TaskProgressUpdate(task, progress, expectedFileSize));
+        processProgressUpdate(
+            TaskProgressUpdate(task, progress, expectedFileSize));
       }
       _retrievedLocallyStoredData = true;
     }
@@ -515,7 +520,8 @@ abstract base class BaseDownloader {
   void _emitProgressUpdate(TaskProgressUpdate update) {
     final task = update.task;
     if (task.providesProgressUpdates) {
-      _updateTaskInDatabase(task, progress: update.progress, expectedFileSize: update.expectedFileSize);
+      _updateTaskInDatabase(task,
+          progress: update.progress, expectedFileSize: update.expectedFileSize);
       final taskProgressCallback = groupProgressCallbacks[task.group];
       if (taskProgressCallback != null) {
         taskProgressCallback(update);
@@ -558,13 +564,13 @@ abstract base class BaseDownloader {
         };
       }
       if (status != TaskStatus.paused) {
-        database
-            .updateRecord(TaskRecord(task, status!, progress!,expectedFileSize, taskException));
+        database.updateRecord(TaskRecord(
+            task, status!, progress!, expectedFileSize, taskException));
       } else {
         // if paused, don't modify the stored progress
         final existingRecord = await database.recordForId(task.taskId);
-        database.updateRecord(TaskRecord(
-            task, status!, existingRecord?.progress ?? 0, expectedFileSize, taskException));
+        database.updateRecord(TaskRecord(task, status!,
+            existingRecord?.progress ?? 0, expectedFileSize, taskException));
       }
     }
   }