+++ lib/src/desktop_downloader_isolate.dart
@@ -430,7 +430,8 @@ Future<TaskStatus> transferBytes(
         if (contentLength > 0 &&
             (progress - lastProgressUpdate > 0.02 &&
                 now.isAfter(nextProgressUpdateTime))) {
-          processProgressUpdateInIsolate(task, progress, sendPort, contentLength);
+          processProgressUpdateInIsolate(
+              task, progress, sendPort, contentLength);
           lastProgressUpdate = progress;
           nextProgressUpdateTime = now.add(const Duration(milliseconds: 500));
         }
@@ -492,7 +493,8 @@ void processStatusUpdateInIsolate(
 ///
 /// Sends progress update via the [sendPort], if requested
 void processProgressUpdateInIsolate(
-    Task task, double progress, SendPort sendPort, [int expectedFileSize = -1]) {
+    Task task, double progress, SendPort sendPort,
+    [int expectedFileSize = -1]) {
   if (task.providesProgressUpdates) {
     sendPort.send(('progressUpdate', progress, expectedFileSize));
   }