+++ lib/src/native_downloader.dart
@@ -64,7 +64,8 @@ final class NativeDownloader extends BaseDownloader {
           processStatusUpdate(TaskStatusUpdate(task, status, exception));
 
         case ('progressUpdate', [double progress, int expectedFileSize]):
-          processProgressUpdate(TaskProgressUpdate(task, progress, expectedFileSize));
+          processProgressUpdate(
+              TaskProgressUpdate(task, progress, expectedFileSize));
 
         case ('canResume', bool canResume):
           setCanResume(task, canResume);